pyproject.toml
[tool.poetry]
name = "connect"
version = "1.0.12"
description = "Connect: weni-engine"
authors = ["weni.ai"]
license = "MPL-2.0"

[tool.poetry.dependencies]
python = "~=3.6.2"
django = "~=3.2.6"
djangorestframework = "~=3.12.4"
drf-yasg2 = "~=1.19.4"
django-environ-2 = "~=2.1.0"
mozilla-django-oidc = "~=2.0.0"
django-filter = "~=2.4.0"
celery = "~=5.1.2"
django-celery-results = "~=2.2.0"
redis = "~=3.5.3"
django-celery-beat = "~=2.2.1"
psycopg2-binary = "~=2.9.1"
python-keycloak = "~=0.25.0"
boto3 = "~=1.18.13"
django-storages = "~=1.11.1"
pillow = "~=8.3.1"
filetype = "~=1.0.7"
django-cors-headers = "~=3.7.0"
whitenoise = "~=5.3.0"
sentry-sdk = "~=1.3.1"
elastic-apm = "~=6.3.3"
django-timezone-field = "~=4.2.1"
grpcio = "~=1.39.0"
grpcio-tools = "~=1.39.0"
djangogrpcframework = "~=0.2.1"
stripe = "2.76.0"
psutil = "~=5.8.0"
elasticsearch = "<7.14"
elasticsearch-dsl = "~=7.4.0"
pendulum = "^2.1.2"
pylint = "^2.13.9"
weni-protobuffers = "1.2.19a0"
freezegun = "^1.2.1"
pre-commit = "2.17"
pandas = "1.0.0"

[tool.poetry.dev-dependencies]
flake8 = "~=3.9.2"
black = "==21.7b0"
coverage = "~=5.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
